src/utils/rulekit/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
idna==2.10
iniconfig==1.1.1
joblib==1.0.0
JPype1==1.2.1
more-itertools==8.8.0
numpy==1.20.0
packaging==20.9
pandas==1.2.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==4.6.11
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
urllib3==1.26.5
wcwidth==0.2.5
